EL-2044: Refactor tests
